Split `ClassFileMethodAccessorGenerator` to better represent the code generated by each "modifier type"
This slightly increases duplication but makes it easier to read overall

Should also help later when generating the same function but forcing without coroutines (for blocking call)
